cors 설정 수정

+++ src/main/java/kr/kro/moonlightmoist/shopapi/coupon/domain/Coupon.java
@@ -88,7 +88,7 @@ public class Coupon extends BaseTimeEntity {
 
     @Column(nullable = true)
     @Builder.Default
-    private Integer issueCount = 0;
+    private Long issueCount = 0L;
 
     @Column(name = "is_deleted", nullable = false)
     @Builder.Default
