+++ src/main/java/kr/kro/moonlightmoist/shopapi/coupon/dto/CouponDto.java
@@ -40,7 +40,7 @@ public class CouponDto {
     @JsonProperty("hasLimitMaxDiscount")
     private Boolean limitMaxDiscountAmount;
     private Integer maxDiscountAmount;
-    private Integer issueCount;
+    private Long issueCount;
     private Boolean deleted;
 
     public Coupon toEntity() {
